Change find/search behavior of projects, sessions, and trials
Now, searching a project on an array of projects will return the matched project (or a list of close-name matches). A search on a single project will return the matching trials.
Consequently, searching on an array of sessions will return the matching session; searching on a single session will return the matching trial
Finally, searching on an array of trials will return the matching trial